+++ packages/web-component/src/lib/descope-wc/DescopeWc.ts
@@ -574,6 +574,9 @@ class DescopeWc extends BaseDescopeWc {
       return;
     }
 
+    sdkResp.data?.runnerLogs?.forEach((l) =>
+      this.loggerWrapper.info(l.title, l.log),
+    );
     const errorText = sdkResp.data?.screen?.state?.errorText;
     if (sdkResp.data?.error) {
       this.loggerWrapper.error(
@@ -692,7 +695,7 @@ class DescopeWc extends BaseDescopeWc {
     }
   }
 
-  #handleDescopePassword(ele: Element) {
+  #handleExternalInputs(ele: Element) {
     if (!ele) {
       return;
     }
@@ -868,6 +871,9 @@ class DescopeWc extends BaseDescopeWc {
       setTimeout(() => {
         updateScreenFromScreenState(this.rootElement, screenState);
 
+        const emailEles = this.rootElement.querySelectorAll(
+          'descope-email-field',
+        );
         const passwordEles =
           this.rootElement.querySelectorAll('descope-password');
         const newPasswordEles = this.rootElement.querySelectorAll(
@@ -879,8 +885,8 @@ class DescopeWc extends BaseDescopeWc {
           .querySelectorAll('[data-hidden-input="true"]')
           .forEach((ele) => ele.remove());
         // handle external input workaround for password components
-        [...passwordEles, ...newPasswordEles].forEach((ele) =>
-          this.#handleDescopePassword(ele),
+        [...emailEles, ...passwordEles, ...newPasswordEles].forEach((ele) =>
+          this.#handleExternalInputs(ele),
         );
       });
 
@@ -934,15 +940,20 @@ class DescopeWc extends BaseDescopeWc {
   }
 
   #validateInputs() {
-    return Array.from(this.shadowRoot.querySelectorAll('*[name]')).every(
-      (input: HTMLInputElement) => {
+    let isValid = true;
+    Array.from(this.shadowRoot.querySelectorAll('*[name]'))
+      .reverse()
+      .forEach((input: HTMLInputElement) => {
         if (input.localName === 'slot') {
-          return true;
+          return;
         }
         input.reportValidity?.();
-        return input.checkValidity?.();
-      },
-    );
+        if (isValid) {
+          isValid = input.checkValidity?.();
+        }
+      });
+
+    return isValid;
   }
 
   async #getFormData() {